created pet model
