--- Modules/Owner/IOwnerService.cs
@@ -1,6 +1,0 @@
-namespace PetCare.Modules.Owner;
-
-public interface IOwnerService
-{
-    public List<Owner> FindByLastName(string lastName);
-}
