+++ Modules/OwnerModule/IOwnerService.cs
@@ -1,0 +1,6 @@
+namespace PetCare.Modules.OwnerModule;
+
+public interface IOwnerService
+{
+    public IEnumerable<Owner> FindByLastName(string lastName);
+}
