+++ Modules/PetModule/Pet.cs
@@ -1,12 +1,12 @@
 using System.ComponentModel.DataAnnotations;
 using PetCare.Models;
 
-namespace PetCare.Pet;
+namespace PetCare.Modules.PetModule;
 
 public class Pet : BaseEntity
 {
     public string? Name { get; set; }
 
     [DataType(DataType.Date)]
-    public DateTime Birthdate { get; set; }
+    public DateTime? Birthdate { get; set; }
 }
